chore: reuse hook
